IBTRS: clt: remove cm_done completion, use single state_wq object
On clt_open or reconnect_work we are on synchronous path, thus detect
errors and close/free by ourselves.  DEVICE_REMOVAL is an async
event, so closing should be taken with care.

Remove cm_done and abuse state_wq as a single wait object for all
events, even for those events (EVENT_ESTABLISHED) which do not change
session state, but set cm_err to 1.