IBTRS: clt,srv: introduce 'recon_cnt' as a reconnection counter
There is a problem, when IBTRS session starts closing and reports
upper layer with SESS_DISCONNECTING.  In case of IBNBD this "disconect"
can take minutes if disks are busy and all bios have to be gracefully
waited.  Meanwhile IBTRS client can start reconnecting and IBTRS
server detects that session is still alive (waiting for inflights).
In this rare case IBTRS server has to reject connection with a special
code -EBUSY and IBTRS client will go to another reconnect iteration.

recon_cnt is needed to distinguish new session and its connections
from session with the same uuid, but which starts reconnecting
procedure.  If server detects that session exists but reconnect
counter is different - -EBUSY is returned.

PS. Now session uuid is generated once on session createion.