ibnbd/ibtrs: pass usr msg explicitely in API
Pass data and user message separately in the API on server side. This
makes it more consistent with the API on client side, where ibnbd passes
data and user message separately to ibtrs.

Also this makes ibnbd independant of the order in which ibtrs trasnsfers
and writes data and user message.

Signed-off-by: Danil Kipnis <danil.kipnis@profitbricks.com>