ibtrs-clt: add check if path to be added already exists
ibtrs spits out a warning into dmesg when trying to add an already
existing path to a session. The waring is generated by sysfs since we
try to create an existing directory.

Check the destination addresses of all paths of the session and
return error if a path with the new address already exists.

Signed-off-by: Danil Kipnis <danil.kipnis@profitbricks.com>