IBNBD: clt: remove get/put sess on session reopen path
It is guaranteed that session will be alive on each IBTRS event,
thus it is totally useless to get/put session on reopen path.
Even session is going to be destroyed and none devices are mapped
nothing terrible will happen if we let ibtrs_clt_sess_reopen()
to be invoked.