IBNBD: clt: make atomic find-get-and-create session routine
There is no any need in mutex lock for the whole period of session
creation, there is no need in session states (will be removed in
future patch), what is really needed is atomic find-get-and-create.

If session is observed with ref == 0 - return -EBUSY.  That is needed
to solve the race when session with particular name is added to the
list, but ibtrs connection is not yet established.  Possible
concurrent access can observe session with 0 reference and will
return -EBUSY.

Of course we can wait, but for now it is not needed.