IBNBD: clt: wait for IBTRS connection for newly created session
In order to keep sessions unique (by names) it is required to put session
into the list *before* IBTRS connection is fully established, thus this
not fully inited session can be observed in the list.

Onother code path which observes session in the list must wait for IBTRS
connection to be fully established.

If error happens session is considered as dying and no attempts to
increase a reference are made.