ibtrs: add min-inflight multipath policy
Selecting the path for an IO in round robin fashion has following problems:

1. If one shuts down a port on an IB switch somewhere on a path, it
takes up to 30 seconds until IB generates an event on the client/server.
We post requests on that dying path until we have full queuedepth of
outstanding requests stuck on that path.
This leads to no IO's at all transferred during that 30 seconds.
2. If one establishes 2 pathes to one server and only 1 path to another
server, maps devices from both of them and starts io, one can see
that one path of the 2-paths-session is idling most of the time, while
the other path is under load.

Introduce the min-inflight policy which chooses the path with minimal number of
outstanding requests. That strategy covers the two usecases mentioned.

Make the policy configurable under
/sys/kernel/ibtrs_client/<sessname>/mp_policy.

Signed-off-by: Danil Kipnis <danil.kipnis@profitbricks.com>
CC: Roman Penyaev <roman.penyaev@profitbricks.com>