IBTRS: clt: wait for graceful shutdown of each conn
In normal scenario (when no errors expected) we sync with server
and wait for RDMA_CM_EVENT_DISCONNECTED event, which is fired
on client side when server explicitly does rdma_disconnect().

That will eliminate possible race when client finishes with
closing of ibtrs connection and immediately establishes new
connection, but server meanwhile is on the way and has not
yet finished with tearing-down connections, sysfs entries
removal, etc.