IBTRS: clt: prepare completion RDMA req from INV callbacks
The idea is simple: in future client will request invalidation
and server responds with IB_WR_SEND_WITH_INV.

But if request is marked for invalidation, but INV has not come
(WHY?) we have to do that ourselves.  Exactly for that reason
the whole stuff in this patch is implemented.