ibtrs: rename ibtrs_tag to ibtrs_permit to avoid confusion with mq tags
Number of inflight IOs in one ibtrs session is limited by the number of
preallocated buffers on the server side of the session. An ibtrs user
needs to prereserve one of those buffers by calling ibtrs_get_tag. This
leads to confusion with the block layer MQ tags if the user of ibtrs is
a block device (IBNBD) exposing one or more blockdevices mapped over a
given ibtrs session.

To avoid confusion use the term "permit" instead of "tag" in ibtrs API.
One ibtrs_permit is a permit to write or request a read for an sg list
on a given ibtrs session.

Signed-off-by: Danil Kipnis <danil.kipnis@cloud.ionos.com>