rtrs-srv: fix kobject warning
We had warning from kobject eg:
kobject: '(null)' (000000002608544b): is not initialized, yet kobject_put() is being called.

The thing is we create sess files in process_info_req, if any error
happens, we queue close_work to remove the sysfs, it's possible
close_work will call kobject_put even kojbect is not initialized.

To fix this add a free_sess function only call kobject_put when it is
initialized.

Signed-off-by: Danil Kipnis <danil.kipnis@cloud.ionos.com>
Signed-off-by: Jack Wang <jinpu.wang@cloud.ionos.com>